Move to own package outside of beacon-chain
