+++ beacon-chain/BUILD.bazel
@@ -79,7 +79,10 @@ docker_push(
 go_binary(
     name = "beacon-chain",
     embed = [":go_default_library"],
-    visibility = ["//beacon-chain:__subpackages__"],
+    visibility = [
+        "//beacon-chain:__subpackages__",
+        "//endtoend:__subpackages__",
+    ],
 )
 
 go_test(
