+++ endtoend/BUILD.bazel
@@ -14,12 +14,13 @@ go_test(
         "@com_github_ethereum_go_ethereum//cmd/geth",
     ],
     tags = [
+        "block-network",
         "exclusive",
         "manual",
         "minimal",
     ],
     deps = [
-        "//beacon-chain/endtoend/evaluators:go_default_library",
+        "//endtoend/evaluators:go_default_library",
         "//contracts/deposit-contract:go_default_library",
         "//proto/eth/v1alpha1:go_default_library",
         "//shared/params:go_default_library",
