+++ endtoend/minimal_e2e_test.go
@@ -3,17 +3,18 @@ package endtoend
 import (
 	"testing"
 
-	ev "github.com/prysmaticlabs/prysm/beacon-chain/endtoend/evaluators"
+	ev "github.com/prysmaticlabs/prysm/endtoend/evaluators"
 	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 )
 
-func TestEndToEnd_DemoConfig(t *testing.T) {
+func TestEndToEnd_MinimalConfig(t *testing.T) {
 	testutil.ResetCache()
-	params.UseDemoBeaconConfig()
-	demoConfig := &end2EndConfig{
-		minimalConfig:  false,
-		epochsToRun:    6,
+	params.UseMinimalConfig()
+
+	minimalConfig := &end2EndConfig{
+		minimalConfig:  true,
+		epochsToRun:    5,
 		numBeaconNodes: 4,
 		numValidators:  params.BeaconConfig().MinGenesisActiveValidatorCount,
 		evaluators: []ev.Evaluator{
@@ -21,5 +22,5 @@ func TestEndToEnd_DemoConfig(t *testing.T) {
 			ev.FinalizationOccurs,
 		},
 	}
-	runEndToEndTest(t, demoConfig)
+	runEndToEndTest(t, minimalConfig)
 }
