+++ validator/BUILD.bazel
@@ -11,9 +11,7 @@ go_library(
         "usage.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/validator",
-    visibility = [
-        "//validator:__subpackages__",
-    ],
+    visibility = ["//validator:__subpackages__"],
     deps = [
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
@@ -82,7 +80,7 @@ go_binary(
     pure = "on",  # Enabled unless there is a valid reason to include cgo dep.
     visibility = [
         "//validator:__subpackages__",
-        "//beacon-chain/endtoend:__pkg__",
+        "//endtoend:__pkg__",
     ],
 )
